@@ -11,11 +11,13 @@
     "axios": "^0.26.1",
     "firebase": "^9.6.10",
     "google-map-react": "^2.1.10",
+    "leaflet": "^1.8.0",
     "moment": "latest",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-leaflet": "^4.0.2",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "redux-thunk": "^2.4.1",
     "sass": "^1.49.9",
