@@ -1,26 +1,35 @@
-import React from "react";
-import GoogleMapReact from "google-map-react";
+import React, { useEffect, useState } from "react";
 
-const AnyReactComponent = ({ text }) => <div>{text}</div>;
+import { MapContainer, TileLayer, Marker, useMap } from "react-leaflet";
+import { getLocation } from "../../services/geolocationApi";
 
-function SimpleMap() {
-  const defaultProps = {
-    center: {
-      lat: 43.6532,
-      lng: 79.3832,
-    },
-    zoom: 9,
-  };
-
+function SimpleMap({ address }) {
+  const [latitude, setLatitude] = useState(null);
+  const [longitude, setLongitude] = useState(null);
+  useEffect(() => {
+    const getGeo = async () => {
+      const [la, lo] = await getLocation(address);
+      setLatitude(la);
+      setLongitude(lo);
+    };
+    getGeo();
+  }, []);
   return (
-    <div className="map-simple" style={{ height: "20vh", width: "100%" }}>
-      <GoogleMapReact
-        bootstrapURLKeys={{ key: "" }}
-        defaultCenter={defaultProps.center}
-        defaultZoom={defaultProps.zoom}
-      >
-        <AnyReactComponent lat={59.955413} lng={30.337844} text="My Marker" />
-      </GoogleMapReact>
+    <div className="map-simple" style={{ height: "30vh", width: "100%" }}>
+      {latitude && longitude && (
+        <MapContainer
+          className="leaflet-container"
+          center={[latitude, longitude]}
+          zoom={11}
+          scrollWheelZoom={false}
+        >
+          <TileLayer
+            attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
+            url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+          />
+          <Marker position={[latitude, longitude]}></Marker>
+        </MapContainer>
+      )}
     </div>
   );
 }
